requirements.txt
flask==3.0.3
flask_sqlalchemy==3.1.1
sqlalchemy==2.0.32
flask-jwt-extended==4.6.0
flask-migrate==4.0.7
alembic==1.13.2
werkzeug==3.0.3
paho-mqtt==1.6.1
apscheduler==3.10.4
python-dotenv==1.0.1
email-validator==2.2.0
marshmallow==3.21.3
flasgger==0.9.7.1
itsdangerous==2.2.0
pytest==8.2.2
requests==2.32.3
waitress==3.0.0
python-dateutil==2.9.0.post0
faker==26.0.0
